requirements.txt
# ===== Core Vector Store =====
chromadb==1.5.1

# ===== Embeddings =====
sentence-transformers==5.2.3
torch==2.10.0
transformers==4.57.6

# ===== Agent Framework =====
langchain==1.2.10
langchain-core==1.2.14
langchain-community==0.4.1
langchain-ollama==1.0.1
langchain-text-splitters==1.1.0
langgraph==1.0.8

# ===== LLM =====
ollama==0.6.1

# ===== Web Search =====
ddgs==9.10.0

# ===== PDF / Extraction =====
PyMuPDF==1.27.1
pymupdf4llm==0.3.4
pdfplumber==0.11.9
pytesseract==0.3.13

# ===== Config =====
python-dotenv==1.2.1
PyYAML==6.0.3

# ===== UI =====
streamlit==1.54.0

# ===== Utilities =====
tqdm==4.67.3